css/selectors/simple-selectors/attribute-selectors/css3-selectors-lang-056.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Selectors: [lang="es"], xml:lang="es" (html)</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/selectors/#attribute-selectors'>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link href="../../../resources/testharness.css" rel="stylesheet" type="text/css">
<meta name='flags' content='HTMLonly'>
<meta name="assert" content="[Exploratory] A [lang='es'] value that matches an identical xml:lang attribute value will NOT produce styling in pages served as HTML.">
<style type='text/css'>
.test div { width: 50px; }
#box[lang='es'] { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" xml:lang="es">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, " ");
</script>

<div id='log'></div>

</body>
</html>
